TRUNK-4917 - Renamed usaged of Mutable to Changeable
